LayoutTests/css3/flexbox/flex-parsing-expected.txt

Test parsing of the CSS flex() function.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testFlex("100px", "width") is "100px"
PASS testFlex("junk", "width") is ""
PASS testFlex("-webkit-flex()", "width") is ""
PASS testFlex("-webkit-flex(1)", "width") is "-webkit-flex(1 0 0px)"
PASS testFlex("-webkit-flex(0)", "width") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(2.4)", "width") is "-webkit-flex(2.4 0 0px)"
PASS testFlex("-webkit-flex(auto)", "width") is "-webkit-flex(1 0 auto)"
PASS testFlex("-webkit-flex(1px)", "width") is "-webkit-flex(1 0 1px)"
PASS testFlex("-webkit-flex(2em)", "width") is "-webkit-flex(1 0 2em)"
PASS testFlex("-webkit-flex(0px)", "width") is "-webkit-flex(1 0 0px)"
PASS testFlex("-webkit-flex(-2)", "width") is ""
PASS testFlex("-webkit-flex(foo)", "width") is ""
PASS testFlex("-webkit-flex(0 0)", "width") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(0 1)", "width") is "-webkit-flex(0 1 0px)"
PASS testFlex("-webkit-flex(1 0)", "width") is "-webkit-flex(1 0 0px)"
PASS testFlex("-webkit-flex(2 auto)", "width") is "-webkit-flex(2 0 auto)"
PASS testFlex("-webkit-flex(3 4px)", "width") is "-webkit-flex(3 0 4px)"
PASS testFlex("-webkit-flex(auto 5.2)", "width") is "-webkit-flex(5.2 0 auto)"
PASS testFlex("-webkit-flex(6em 4)", "width") is "-webkit-flex(4 0 6em)"
PASS testFlex("-webkit-flex(4 0px)", "width") is "-webkit-flex(4 0 0px)"
PASS testFlex("-webkit-flex(0 0px)", "width") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(-1 5)", "width") is ""
PASS testFlex("-webkit-flex(1 -1)", "width") is ""
PASS testFlex("-webkit-flex(-1 -1)", "width") is ""
PASS testFlex("-webkit-flex(auto 2em)", "width") is ""
PASS testFlex("-webkit-flex(2px 4px)", "width") is ""
PASS testFlex("-webkit-flex(0px 0px)", "width") is ""
PASS testFlex("-webkit-flex(0 0 0)", "width") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(1 2 0)", "width") is "-webkit-flex(1 2 0px)"
PASS testFlex("-webkit-flex(1 2 auto)", "width") is "-webkit-flex(1 2 auto)"
PASS testFlex("-webkit-flex(1.6 2 3px)", "width") is "-webkit-flex(1.6 2 3px)"
PASS testFlex("-webkit-flex(1 3px 2)", "width") is "-webkit-flex(1 2 3px)"
PASS testFlex("-webkit-flex(3px 1 2)", "width") is "-webkit-flex(1 2 3px)"
PASS testFlex("-webkit-flex(auto 0 0)", "width") is "-webkit-flex(0 0 auto)"
PASS testFlex("-webkit-flex(0 0px 0)", "width") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(1 2 3)", "width") is ""
PASS testFlex("-webkit-flex(0 2 3)", "width") is ""
PASS testFlex("-webkit-flex(1 0 3)", "width") is ""
PASS testFlex("-webkit-flex(0 0 1)", "width") is ""
PASS testFlex("-webkit-flex(1 -2 3px)", "width") is ""
PASS testFlex("-webkit-flex(1 2px 3px)", "width") is ""
PASS testFlex("-webkit-flex(1 2px auto)", "width") is ""
PASS testFlex("-webkit-flex(0px 0px 0)", "width") is ""
PASS testFlex("-webkit-flex(0 0 0 0)", "width") is ""
PASS testFlex("-webkit-flex(0 0 0px 0)", "width") is ""
PASS testFlex("-webkit-flex(0 0 0px 0px)", "width") is ""
PASS testFlex("-webkit-flex(1) 10px", "width") is ""
PASS testFlex("-webkit-flex(1) auto", "width") is ""
PASS testFlex("-webkit-flex(1) foo", "width") is ""
PASS testFlex("10px -webkit-flex(1)", "width") is ""
PASS testFlex("100px", "height") is "100px"
PASS testFlex("junk", "height") is ""
PASS testFlex("-webkit-flex()", "height") is ""
PASS testFlex("-webkit-flex(1)", "height") is "-webkit-flex(1 0 0px)"
PASS testFlex("-webkit-flex(0)", "height") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(2.4)", "height") is "-webkit-flex(2.4 0 0px)"
PASS testFlex("-webkit-flex(auto)", "height") is "-webkit-flex(1 0 auto)"
PASS testFlex("-webkit-flex(1px)", "height") is "-webkit-flex(1 0 1px)"
PASS testFlex("-webkit-flex(2em)", "height") is "-webkit-flex(1 0 2em)"
PASS testFlex("-webkit-flex(0px)", "height") is "-webkit-flex(1 0 0px)"
PASS testFlex("-webkit-flex(-2)", "height") is ""
PASS testFlex("-webkit-flex(foo)", "height") is ""
PASS testFlex("-webkit-flex(0 0)", "height") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(0 1)", "height") is "-webkit-flex(0 1 0px)"
PASS testFlex("-webkit-flex(1 0)", "height") is "-webkit-flex(1 0 0px)"
PASS testFlex("-webkit-flex(2 auto)", "height") is "-webkit-flex(2 0 auto)"
PASS testFlex("-webkit-flex(3 4px)", "height") is "-webkit-flex(3 0 4px)"
PASS testFlex("-webkit-flex(auto 5.2)", "height") is "-webkit-flex(5.2 0 auto)"
PASS testFlex("-webkit-flex(6em 4)", "height") is "-webkit-flex(4 0 6em)"
PASS testFlex("-webkit-flex(4 0px)", "height") is "-webkit-flex(4 0 0px)"
PASS testFlex("-webkit-flex(0 0px)", "height") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(-1 5)", "height") is ""
PASS testFlex("-webkit-flex(1 -1)", "height") is ""
PASS testFlex("-webkit-flex(-1 -1)", "height") is ""
PASS testFlex("-webkit-flex(auto 2em)", "height") is ""
PASS testFlex("-webkit-flex(2px 4px)", "height") is ""
PASS testFlex("-webkit-flex(0px 0px)", "height") is ""
PASS testFlex("-webkit-flex(0 0 0)", "height") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(1 2 0)", "height") is "-webkit-flex(1 2 0px)"
PASS testFlex("-webkit-flex(1 2 auto)", "height") is "-webkit-flex(1 2 auto)"
PASS testFlex("-webkit-flex(1.6 2 3px)", "height") is "-webkit-flex(1.6 2 3px)"
PASS testFlex("-webkit-flex(1 3px 2)", "height") is "-webkit-flex(1 2 3px)"
PASS testFlex("-webkit-flex(3px 1 2)", "height") is "-webkit-flex(1 2 3px)"
PASS testFlex("-webkit-flex(auto 0 0)", "height") is "-webkit-flex(0 0 auto)"
PASS testFlex("-webkit-flex(0 0px 0)", "height") is "-webkit-flex(0 0 0px)"
PASS testFlex("-webkit-flex(1 2 3)", "height") is ""
PASS testFlex("-webkit-flex(0 2 3)", "height") is ""
PASS testFlex("-webkit-flex(1 0 3)", "height") is ""
PASS testFlex("-webkit-flex(0 0 1)", "height") is ""
PASS testFlex("-webkit-flex(1 -2 3px)", "height") is ""
PASS testFlex("-webkit-flex(1 2px 3px)", "height") is ""
PASS testFlex("-webkit-flex(1 2px auto)", "height") is ""
PASS testFlex("-webkit-flex(0px 0px 0)", "height") is ""
PASS testFlex("-webkit-flex(0 0 0 0)", "height") is ""
PASS testFlex("-webkit-flex(0 0 0px 0)", "height") is ""
PASS testFlex("-webkit-flex(0 0 0px 0px)", "height") is ""
PASS testFlex("-webkit-flex(1) 10px", "height") is ""
PASS testFlex("-webkit-flex(1) auto", "height") is ""
PASS testFlex("-webkit-flex(1) foo", "height") is ""
PASS testFlex("10px -webkit-flex(1)", "height") is ""
PASS successfullyParsed is true

TEST COMPLETE
